requirements.txt
flask>=2.0.0
python-dotenv
openai
langchain
pandas
gunicorn